Fixed connection to SteamVR preventing HMD detection
- Thanks to Kadarus for the help!

+++ Source/2_Core/Managers/SteamVR/SteamVRWebConsoleSession.cs
@@ -22,7 +22,7 @@ namespace BeatLeader.SteamVR {
         }
 
         public struct Message {
-            public int nTimestamp;
+            public long nTimestamp;
             public string sLogLevel;
             public string sLogName;
             public string sMessage;
